test-requirements.txt
numpy
scipy
matplotlib
progressbar2
netCDF4
xarray
dask
toolz
